React router overhaul
I did a sort of overhaul with a lot here. The primary goal which I achieved was adding my components to route to make it easy to see all the bite sized projects I used to facilitate my learning of react. Nothing here is serious and should be taken seriously thx